Integer values for amounts to simplify calculations

+++ Assets/Src/Classes/Entities/PileXML.cs
@@ -16,7 +16,7 @@ public class PileXML
 	public String Name;
 	
 	[XmlAttribute("Quantity")]
-	public float Quantity;
+	public int Quantity;
 }
